ÄNDERUNGEN

+++ Routenplaner/main/de/dhbw/horb/routePlanner/evaluation/dijkstra/Way.java
@@ -12,7 +12,7 @@ import java.util.Map;
  */
 public class Way {
 
-    private Long price = (long) 0;
+    private Double price = (double) 0;
     private List<String> nodes = new ArrayList<String>();
     private List<Map<String, String>> edges = new ArrayList<Map<String, String>>();
 
@@ -20,7 +20,7 @@ public class Way {
 	nodes.add(startnode);
     }
 
-    public Way(List<String> nodes, Long gonePrice, String newNode, Long newPrice, List<Map<String, String>> edges,
+    public Way(List<String> nodes, Double gonePrice, String newNode, Double newPrice, List<Map<String, String>> edges,
 	    Map<String, String> newEdge) {
 	setPrice(gonePrice + newPrice);
 	addNodes(nodes);
@@ -47,11 +47,11 @@ public class Way {
 	return nodes.get(nodes.size() - 1);
     }
 
-    public Long getPrice() {
+    public Double getPrice() {
 	return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(Double price) {
 	this.price = price;
     }
